# service/admin/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/orders": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get orders with pagination and filters. Only for ADMINs.",
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Completed. Complteted must be 'true', 'false' or ''",
                        "name": "completed",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succesfull get of orders",
                        "schema": {
                            "$ref": "#/definitions/dto.Orders"
                        }
                    },
                    "400": {
                        "description": "Id must be uuid",
                        "schema": {
                            "$ref": "#/definitions/resp.JsonError"
                        }
                    },
                    "401": {
                        "description": "Unauth",
                        "schema": {
                            "$ref": "#/definitions/resp.JsonError"
                        }
                    },
                    "403": {
                        "description": "Invalid role",
                        "schema": {
                            "$ref": "#/definitions/resp.JsonError"
                        }
                    },
                    "500": {
                        "description": "Something going wrong...",
                        "schema": {
                            "$ref": "#/definitions/resp.JsonError"
                        }
                    }
                }
            }
        },
        "/admin/orders/stats": {
            "get": {
                "description": "Get stats of order creations. Only for ADMINs.",
                "tags": [
                    "Orders"
                ],
                "summary": "Get stats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter",
                        "name": "filter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Stats"
                        }
                    },
                    "500": {
                        "description": "Something going wrong...",
                        "schema": {
                            "$ref": "#/definitions/resp.JsonError"
                        }
                    }
                }
            }
        },
        "/admin/orders/{id}": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Set order. Only for ADMINs",
                "tags": [
                    "Orders"
                ],
                "summary": "Set order completed",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful set order status"
                    },
                    "400": {
                        "description": "Id must be uuid",
                        "schema": {
                            "$ref": "#/definitions/resp.JsonError"
                        }
                    },
                    "500": {
                        "description": "Something going wrong...",
                        "schema": {
                            "$ref": "#/definitions/resp.JsonError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.Order": {
            "type": "object",
            "properties": {
                "booking_id": {
                    "type": "string"
                },
                "booking_title": {
                    "type": "string"
                },
                "completed": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "thing": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.Orders": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "orders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Order"
                    }
                }
            }
        },
        "dto.Stats": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
...
        "resp.JsonError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}